requirements.txt
django~=4.1
djangorestframework~=3.14
psycopg2-binary~=2.9
django-crispy-forms~=1.14
django-sortedm2m~=3.1
requests~=2.28
pyyaml~=6.0
django-redis~=5.2.0
drf-spectacular~=0.25
transliterate~=1.10
django-debug-toolbar
gunicorn
